Move quiz state to common module
